requirements.txt
fastapi==0.79.0
ffmpeg-python
ginza==5.1.2
google-cloud-storage==2.5.0
ja-ginza-electra==5.1.2
python-multipart==0.0.5
redis==4.3.4
spacy==3.4.1
torch==1.12.1
torchvision==0.13.1
ulid-py==1.1.0
uvicorn==0.18.2
